pyproject.toml
[project]
name = "llm-epistemic-belief-7a32"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.11.0",
    "datasets>=4.4.0",
    "matplotlib>=3.10.7",
    "numpy>=2.3.4",
    "pandas>=2.3.3",
    "python-docx>=1.2.0",
    "scikit-learn>=1.7.2",
    "scipy>=1.16.3",
    "seaborn>=0.13.2",
    "sentencepiece>=0.2.1",
    "torch>=2.9.0",
    "tqdm>=4.67.1",
    "transformers>=4.57.1",
]
